Int identifiers chanded to decimal because of Oracle field type system

+++ PoliticPolls.Web/Controllers/OrdersController.cs
@@ -23,7 +23,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Orders/Details/5
-        public ActionResult Details(int? id)
+        public ActionResult Details(decimal? id)
         {
             if (id == null)
             {
@@ -63,7 +63,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Orders/Edit/5
-        public ActionResult Edit(int? id)
+        public ActionResult Edit(decimal? id)
         {
             if (id == null)
             {
@@ -96,7 +96,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Orders/Delete/5
-        public ActionResult Delete(int? id)
+        public ActionResult Delete(decimal? id)
         {
             if (id == null)
             {
@@ -113,7 +113,7 @@ namespace PoliticPolls.Web.Controllers
         // POST: Orders/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(decimal id)
         {
             var orders = db.Orders.Find(id);
             db.Orders.Remove(orders);
