+++ PoliticPolls.Web/Controllers/PoliticiansController.cs
@@ -26,7 +26,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Politicians/Details/5
-        public ActionResult Details(int? id)
+        public ActionResult Details(decimal? id)
         {
             if (id == null)
             {
@@ -66,7 +66,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Politicians/Edit/5
-        public ActionResult Edit(int? id)
+        public ActionResult Edit(decimal? id)
         {
             if (id == null)
             {
@@ -99,7 +99,7 @@ namespace PoliticPolls.Web.Controllers
         }
 
         // GET: Politicians/Delete/5
-        public ActionResult Delete(int? id)
+        public ActionResult Delete(decimal? id)
         {
             if (id == null)
             {
@@ -116,7 +116,7 @@ namespace PoliticPolls.Web.Controllers
         // POST: Politicians/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(decimal id)
         {
             var politicians = db.Politicians.Find(id);
             db.Politicians.Remove(politicians);
